dump1090: return uint16_t from getHeading to match the field
heading is a uint16_t member with a uint16_t setter, but getHeading()
returned uint32_t — the only getter whose return type didn't match its
field. Harmless (heading <= 360) but inconsistent with getSquawk()/
getVertVel(); align it.

+++ src/dump1090.hpp
@@ -53,7 +53,7 @@ public:
         this->heading = t_heading;
         this->heading_set = true;
     };
-    auto getHeading() -> uint32_t { return this->heading; };
+    auto getHeading() -> uint16_t { return this->heading; };
     auto validHeading() -> bool { return this->heading_set; };
     void setVertVel(int16_t t_vert_vel)
     {
